PPT/code/refractor/src/refractor/Rental.java
package refractor;

public class Rental {
	private Movie _movie;
	private int _daysRented;
	
	public Rental(Movie movie, int daysRented) {
		_movie = movie;
		_daysRented = daysRented;
	}
	
	public int getDaysRented() {
		return _daysRented;
	}
	
	public Movie getMovie() {
		return _movie;
	}
	
	public double getCharge() {
		return _movie.getCharge(_daysRented);
	}
	
	public int getFrequentRenterPoints() {
		//访问频次加1		
		//新发行电影租借2天以上再加1
		return _movie.getFrequentRenterPoints(_daysRented);
	}
}
